Implement camera fallback for iOS simulator
