+++ sample/iosApp/iosApp/iOSApp.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import ComposeApp
 
 @main
 struct iOSApp: App {
@@ -7,4 +8,8 @@ struct iOSApp: App {
 			ContentView()
 		}
 	}
+    
+    init() {
+        WebRtc.shared.configurePeerConnectionFactory(loggingSeverity: .rtcloggingseveritywarning)
+    }
 }
